Convert target tiles to SO
