+++ Assets/Persistent Data/ClassSOs/Enemy/BanditArcher/BowShot.asset
@@ -33,4 +33,4 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 5594ad7acc2dc44508b9aaf0cae4c662, type: 2}
   - {fileID: 11400000, guid: 9cceed499ff45954c93f9eff929550ab, type: 2}
   - {fileID: 11400000, guid: 8f8423bc60ccc4a4a95c3efc1f66ed53, type: 2}
-  m_TargetSquares: []
+  m_TargetSO: {fileID: 11400000, guid: 8a24e54c7d084402bbec2086c58fc80e, type: 2}
