+++ Assets/Persistent Data/ClassSOs/Enemy/BanditSwordsman/Slash.asset
@@ -33,4 +33,4 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 5594ad7acc2dc44508b9aaf0cae4c662, type: 2}
   - {fileID: 11400000, guid: bafa5c363338c0849bc028b7798a2cd0, type: 2}
   - {fileID: 11400000, guid: 3532326be7568d1419b024ef2dccee87, type: 2}
-  m_TargetSquares: []
+  m_TargetSO: {fileID: 11400000, guid: 8a24e54c7d084402bbec2086c58fc80e, type: 2}
