+++ Assets/Persistent Data/ClassSOs/Player/Archer/TacticalRetreat.asset
@@ -42,4 +42,4 @@ MonoBehaviour:
   m_SkillAnimationType: 1
   m_TargetRules:
   - {fileID: 11400000, guid: d268b682d0c8745928210e76eab4a622, type: 2}
-  m_TargetSquares: []
+  m_TargetSO: {fileID: 11400000, guid: 8a24e54c7d084402bbec2086c58fc80e, type: 2}
