+++ Assets/Persistent Data/ClassSOs/Player/Archer/Volley.asset
@@ -34,20 +34,4 @@ MonoBehaviour:
   - {fileID: 11400000, guid: 9cceed499ff45954c93f9eff929550ab, type: 2}
   - {fileID: 11400000, guid: 8f8423bc60ccc4a4a95c3efc1f66ed53, type: 2}
   - {fileID: 11400000, guid: 5594ad7acc2dc44508b9aaf0cae4c662, type: 2}
-  m_TargetSquares:
-  - m_Row: -1
-    m_Col: -1
-  - m_Row: -1
-    m_Col: 0
-  - m_Row: -1
-    m_Col: 1
-  - m_Row: 0
-    m_Col: -1
-  - m_Row: 0
-    m_Col: 1
-  - m_Row: 1
-    m_Col: -1
-  - m_Row: 1
-    m_Col: 0
-  - m_Row: 1
-    m_Col: 1
+  m_TargetSO: {fileID: 11400000, guid: 07ceb5a02e5a44824bae1668e0b579bf, type: 2}
