+++ Assets/Persistent Data/ClassSOs/Player/Lord/Rally.asset
@@ -39,20 +39,4 @@ MonoBehaviour:
   m_SkillAnimationType: 1
   m_TargetRules:
   - {fileID: 11400000, guid: d268b682d0c8745928210e76eab4a622, type: 2}
-  m_TargetSquares:
-  - m_Row: -1
-    m_Col: -1
-  - m_Row: -1
-    m_Col: 0
-  - m_Row: -1
-    m_Col: 1
-  - m_Row: 0
-    m_Col: -1
-  - m_Row: 0
-    m_Col: 1
-  - m_Row: 1
-    m_Col: -1
-  - m_Row: 1
-    m_Col: 0
-  - m_Row: 1
-    m_Col: 1
+  m_TargetSO: {fileID: 11400000, guid: 07ceb5a02e5a44824bae1668e0b579bf, type: 2}
