+++ Assets/Persistent Data/ClassSOs/Player/Lord/Slash.asset
@@ -32,4 +32,4 @@ MonoBehaviour:
   m_TargetRules:
   - {fileID: 11400000, guid: bafa5c363338c0849bc028b7798a2cd0, type: 2}
   - {fileID: 11400000, guid: 5594ad7acc2dc44508b9aaf0cae4c662, type: 2}
-  m_TargetSquares: []
+  m_TargetSO: {fileID: 11400000, guid: 8a24e54c7d084402bbec2086c58fc80e, type: 2}
