+++ Assets/Persistent Data/ClassSOs/Player/Lord/Sweep.asset
@@ -32,8 +32,4 @@ MonoBehaviour:
   m_TargetRules:
   - {fileID: 11400000, guid: 5594ad7acc2dc44508b9aaf0cae4c662, type: 2}
   - {fileID: 11400000, guid: 3532326be7568d1419b024ef2dccee87, type: 2}
-  m_TargetSquares:
-  - m_Row: 0
-    m_Col: -1
-  - m_Row: 0
-    m_Col: 1
+  m_TargetSO: {fileID: 11400000, guid: b5c0f3b9ecd2f4d1da499bf0f06116a8, type: 2}
