+++ Assets/Persistent Data/ClassSOs/Player/Mage/Fireball.asset
@@ -35,4 +35,4 @@ MonoBehaviour:
   m_SkillAnimationType: 2
   m_TargetRules:
   - {fileID: 11400000, guid: 5594ad7acc2dc44508b9aaf0cae4c662, type: 2}
-  m_TargetSquares: []
+  m_TargetSO: {fileID: 11400000, guid: 8a24e54c7d084402bbec2086c58fc80e, type: 2}
