+++ Assets/Persistent Data/ClassSOs/Player/Mage/Thunderstorm.asset
@@ -31,20 +31,4 @@ MonoBehaviour:
   m_SkillAnimationType: 2
   m_TargetRules:
   - {fileID: 11400000, guid: 5594ad7acc2dc44508b9aaf0cae4c662, type: 2}
-  m_TargetSquares:
-  - m_Row: -1
-    m_Col: -1
-  - m_Row: -1
-    m_Col: 0
-  - m_Row: -1
-    m_Col: 1
-  - m_Row: 0
-    m_Col: -1
-  - m_Row: 0
-    m_Col: 1
-  - m_Row: 1
-    m_Col: -1
-  - m_Row: 1
-    m_Col: 0
-  - m_Row: 1
-    m_Col: 1
+  m_TargetSO: {fileID: 11400000, guid: 07ceb5a02e5a44824bae1668e0b579bf, type: 2}
