+++ Assets/Persistent Data/ClassSOs/Player/Paladin/Blessing.asset
@@ -46,4 +46,4 @@ MonoBehaviour:
   m_TargetRules:
   - {fileID: 11400000, guid: 3c85fb4aacd7243b7ac57ed4a8c4e74b, type: 2}
   - {fileID: 11400000, guid: d268b682d0c8745928210e76eab4a622, type: 2}
-  m_TargetSquares: []
+  m_TargetSO: {fileID: 11400000, guid: 8a24e54c7d084402bbec2086c58fc80e, type: 2}
