+++ Assets/Persistent Data/ClassSOs/Player/Paladin/HandOfLight.asset
@@ -31,4 +31,4 @@ MonoBehaviour:
   m_SkillAnimationType: 1
   m_TargetRules:
   - {fileID: 11400000, guid: 3c85fb4aacd7243b7ac57ed4a8c4e74b, type: 2}
-  m_TargetSquares: []
+  m_TargetSO: {fileID: 11400000, guid: 8a24e54c7d084402bbec2086c58fc80e, type: 2}
