+++ Assets/Persistent Data/ClassSOs/Player/Paladin/Lunge.asset
@@ -32,6 +32,4 @@ MonoBehaviour:
   m_TargetRules:
   - {fileID: 11400000, guid: 5594ad7acc2dc44508b9aaf0cae4c662, type: 2}
   - {fileID: 11400000, guid: 3532326be7568d1419b024ef2dccee87, type: 2}
-  m_TargetSquares:
-  - m_Row: 1
-    m_Col: 0
+  m_TargetSO: {fileID: 11400000, guid: a1be98955ebb4419b91dbb8fe43d80ac, type: 2}
