+++ Assets/Persistent Data/TargetSO.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0793d189bb3654022bfb4dbc84bc3ff3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
