+++ Assets/Persistent Data/TargetSO/1x3 Row Target.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 63bed944bf1a048859d430bac902d08e, type: 3}
+  m_Name: 1x3 Row Target
+  m_EditorClassIdentifier: 
+  m_TargetTiles:
+  - m_Row: 0
+    m_Col: -1
+  - m_Row: 0
+    m_Col: 1
+  m_TargetRepSprite: {fileID: 0}
