+++ Assets/Persistent Data/TargetSO/1x3 Row Target.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5c0f3b9ecd2f4d1da499bf0f06116a8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
