+++ Assets/Persistent Data/TargetSO/2x1 Col Target.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a1be98955ebb4419b91dbb8fe43d80ac
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
