+++ Assets/Persistent Data/TargetSO/SingleTargetSO.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8a24e54c7d084402bbec2086c58fc80e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
