+++ Assets/Scripts/Persistent Data/Active Skills/ActiveSkillSO.cs
@@ -72,10 +72,10 @@ public class ActiveSkillSO : ScriptableObject
 
     [Header("Target")]
     [Tooltip("These are tiles that will also be targeted, represented as offsets from the target square")]
-    public List<CoordPair> m_TargetSquares;
+    public TargetSO m_TargetSO;
 
     #region Helpers
-    public bool IsAoe => m_TargetSquares.Count > 0;
+    public bool IsAoe => m_TargetSO.IsAoe;
     public bool DealsDamage => ContainsSkillType(SkillEffectType.DEALS_DAMAGE);
     public bool IsHeal => ContainsSkillType(SkillEffectType.HEAL);
     public bool IsMagic => m_SkillType == SkillType.MAGIC;
@@ -137,38 +137,11 @@ public class ActiveSkillSO : ScriptableObject
         return m_TeleportTargetRules.All(x => x.IsValidTargetTile(targetTile, unit, targetGridType));
     }
 
-    public List<CoordPair> ConstructAttackTargetTiles(CoordPair target)
-    {
-        List<CoordPair> attackTargetTiles = new() {target};
-
-        foreach (CoordPair offset in m_TargetSquares)
-        {
-            attackTargetTiles.Add(target.Offset(offset));
-        }
-
-        return attackTargetTiles;
-    }
+    public List<CoordPair> ConstructAttackTargetTiles(CoordPair target) => m_TargetSO.ConstructAttackTargetTiles(target);
 
 #if UNITY_EDITOR
     private void OnValidate()
     {
-        HashSet<CoordPair> previousTargets = new();
-        foreach (CoordPair coordPair in m_TargetSquares)
-        {
-            if (coordPair.Equals(new CoordPair(0, 0)))
-            {
-                Logger.Log(this.GetType().Name, $"Target tiles for {name} repeats origin", LogLevel.WARNING);
-            }
-            else if (previousTargets.Contains(coordPair))
-            {
-                Logger.Log(this.GetType().Name, $"Repeated target tile: {coordPair} for {name}", LogLevel.WARNING);
-            }
-            else
-            {
-                previousTargets.Add(coordPair);
-            }
-        }
-
         foreach (InflictedToken inflictedToken in m_InflictedTokens)
         {
             if (inflictedToken == null)
