+++ Assets/Scripts/Persistent Data/Active Skills/TargetSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 63bed944bf1a048859d430bac902d08e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
